pyproject.toml
[project]
name = "mcp-a2a-acp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk[http-server]>=0.3.16",
    "fastapi>=0.121.3",
    "httpx>=0.28.1",
    "mcp>=1.22.0",
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
    "uvicorn==0.31.1",
    "websockets>=15.0.1",
]
